backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt<4.0.0
python-multipart>=0.0.9
google-genai>=1.0
Pillow>=10.3
aiofiles>=23.2
